moved logic from library loader to loader
